Conflicts:
	Makefile
	drivers/Arduino/firmware/arduinoDAC/arduinoDAC.pde

+++ openplacosautoinstall-ubuntu.sh
@@ -44,7 +44,12 @@ installation() {
   cp -rf $path/../openplacos/ /usr/lib/ruby/
 
 # default config
+if [ -e /etc/default/openplacos ]
+then
+  echo "config file already exist"
+else  
   cp $path/server/config_with_VirtualPlacos.yaml /etc/default/openplacos
+fi 
 }
 
 mysql_install() {
@@ -60,6 +65,7 @@ mysql_install() {
 post_installation() {
   # User openplacos
   adduser openplacos --system -disabled-login -no-create-home
+  adduser openplacos dialout
 
 
 # link into path
